backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
dowhy
nashpy
torch
tensorflow
celery
sqlalchemy
alembic
pytest
pydantic
python-multipart
aiofiles
httpx
redis
pinecone-client
neo4j
web3
eth-abi
polygon-api-client
huggingface-hub
transformers
diffusers
deepface
opencv-python
elevenlabs
moviepy
openai
anthropic